started refacto
