+++ go.mod
@@ -1,12 +1,3 @@
 module github.com/Obito1903/CY-celcat
 
-go 1.16
-
-require (
-	github.com/ajstarks/svgo v0.0.0-20200725142600-7a3c8b57fecb
-	github.com/arran4/golang-ical v0.0.0-20210825232153-efac1f4cb8ac
-	github.com/srwiley/oksvg v0.0.0-20210320200257-875f767ac39a
-	github.com/srwiley/rasterx v0.0.0-20200120212402-85cb7272f5e9
-	golang.org/x/image v0.0.0-20210220032944-ac19c3e999fb // indirect
-	golang.org/x/net v0.0.0-20210917221730-978cfadd31cf
-)
+go 1.17
