+++ pkg/celcat/celcat.go
@@ -1,0 +1,9 @@
+package celcat
+
+import "net/url"
+
+type SessionData struct {
+	token         string
+	federationIds string
+	location      url.URL
+}
